Improved Grapling ORB
It is now better, faster and you can stick to blocks and unstick and cancel graple any time with shift, and keep momentum.

Now grapling orb can be used to push people off and if you press shift it applies extra pull so the rope like breaks and throughs the victim player over you.

There is a bunch of coments so any of developers could tweak it later to improve it